added classifier layer
