+++ dataquality/integrations/transformers_trainer.py
@@ -1,81 +1,28 @@
-import gc
-from functools import wraps
-from typing import Any, Callable, Dict, Iterable, Iterator, List, Optional, Union
+# Imports for the hook manager
+from typing import Callable, Dict, Optional
 
 from datasets import Dataset
 from torch.nn import Module
-from torch.utils.data.dataloader import _BaseDataLoaderIter
-from torch.utils.data.sampler import Sampler
 from transformers import Trainer
 from transformers.trainer_callback import TrainerCallback, TrainerControl, TrainerState
 from transformers.training_args import TrainingArguments
 
 import dataquality as dq
+from dataquality.analytics import Analytics
+from dataquality.clients.api import ApiClient
 from dataquality.exceptions import GalileoException
 from dataquality.integrations.torch import TorchBaseInstance
 from dataquality.schemas.split import Split
 from dataquality.schemas.torch import DimensionSlice, InputDim, Layer
+from dataquality.utils.helpers import check_noop
 from dataquality.utils.torch import ModelHookManager
 from dataquality.utils.transformers import (
     add_id_to_signature_columns,
     remove_id_collate_fn_wrapper,
 )
 
-
-def patch_dataloaders(store: Dict, reset_indices: bool = True) -> None:
-    def wrap_next_index(func: Callable, key: str = "ids") -> Callable:
-        @wraps(func)
-        def patched_next_index(*args: Any, **kwargs: Any) -> Any:
-            print("patched_next_index", args, kwargs)
-            indices = func(*args, **kwargs)
-            if indices and key in store:
-                if reset_indices and store.get("last_action") == "pop":
-                    store[key] = []
-                store[key].append(indices)
-                store["last_action"] = "append"
-
-            return indices
-
-        return patched_next_index
-
-    if hasattr(_BaseDataLoaderIter, "_patched"):
-        # logger warning if already patched
-        print("BaseDataLoaderIter already patched")
-        return
-    if "patches" not in store:
-        store["patches"] = []
-
-    store["patches"].append({"class": _BaseDataLoaderIter, "attr": "_next_index"})
-    setattr(_BaseDataLoaderIter, "_old__next_index", _BaseDataLoaderIter._next_index)
-    setattr(
-        _BaseDataLoaderIter,
-        "_next_index",
-        wrap_next_index(_BaseDataLoaderIter._next_index, "ids"),
-    )
-    setattr(_BaseDataLoaderIter, "_patched", True)
-
-
-def unpatch(store: Dict) -> None:
-    # unpatch all instances and classes
-    # starting with all classes
-    for patch in store.get("patches", []):
-        print("unpatching", patch["class"])
-        if hasattr(patch["class"], "_patched"):
-            print("unpatching", patch["class"], patch["attr"])
-            setattr(
-                patch["class"],
-                patch["attr"],
-                getattr(patch["class"], f"_old_{patch['attr']}"),
-            )
-            delattr(patch["class"], f"_old_{patch['attr']}")
-            delattr(patch["class"], "_patched")
-        # then all instances
-        for obj in gc.get_objects():
-            if isinstance(obj, patch["class"]) and hasattr(obj, "_patched"):
-                print("unpatching", obj, patch["attr"])
-                setattr(obj, patch["attr"], getattr(obj, f"old_{patch['attr']}"))
-                delattr(obj, f"old_{patch['attr']}")
-                delattr(obj, "_patched")
+a = Analytics(ApiClient, dq.config)  # type: ignore
+a.log_import("transformers_trainer")
 
 
 class DQCallback(TrainerCallback, TorchBaseInstance):
@@ -90,21 +37,23 @@ class DQCallback(TrainerCallback, TorchBaseInstance):
 
     def __init__(
         self,
-        classifier_layer: Layer = None,
+        last_hidden_state_layer: Layer = None,
         embedding_dim: InputDim = None,
         logits_dim: InputDim = None,
-        embedding_fn: Optional[Any] = None,
-        logits_fn: Optional[Any] = None,
+        classifier_layer: Layer = "classifier",
+        embedding_fn: Optional[Callable] = None,
+        logits_fn: Optional[Callable] = None,
     ) -> None:
         # Access the dq logger helper data
         self.helper_data = dq.get_model_logger().logger_config.helper_data
         self._initialized = False
         # Hook manager for attaching hooks to the model
         self.hook_manager = ModelHookManager()
+        self.last_hidden_state_layer = last_hidden_state_layer
         self.classifier_layer = classifier_layer
-        self._set_dimensions(embedding_dim, logits_dim)
         self.embedding_fn = embedding_fn
         self.logits_fn = logits_fn
+        self._init_dimension(embedding_dim, logits_dim)
 
     def _clear_logger_config_helper_data(self) -> None:
         self.helper_data.clear()
@@ -159,7 +108,9 @@ class DQCallback(TrainerCallback, TorchBaseInstance):
         self.task = dq.config.task_type
         model: Module = kwargs["model"]
         # Attach hooks to the model
-        self._attach_hooks_to_model(model, self.classifier_layer)
+        self._attach_hooks_to_model(
+            model, self.classifier_layer, self.last_hidden_state_layer
+        )
         train_dataloader = kwargs["train_dataloader"]
         train_dataloader_ds = train_dataloader.dataset
         if isinstance(train_dataloader_ds, Dataset):
@@ -260,119 +211,67 @@ class DQCallback(TrainerCallback, TorchBaseInstance):
         """
         self._do_log()
 
-    def _attach_hooks_to_model(self, model: Module, layer: Layer) -> None:
+    def _attach_hooks_to_model(
+        self, model: Module, classifier_layer: Layer, last_hidden_state_layer: Layer
+    ) -> None:
         """
         Method to attach hooks to the model by using the hook manager
         :param model: Model
         :param model: pytorch model layer to attach hooks to
         :return: None
         """
+        try:
+            self.hook_manager.attach_classifier_hook(
+                model, self._classifier_hook, classifier_layer
+            )
+        except Exception as e:
+            print(
+                f"Could not attach classifier hook to model. "
+                f"Error: {e}. "
+                f"Please check the classifier layer name: {classifier_layer}"
+            )
+            self.hook_manager.attach_hooks_to_model(
+                model, self._embedding_hook, last_hidden_state_layer
+            )
+            self.hook_manager.attach_hook(model, self._logit_hook)
 
-        self.hook_manager.attach_embedding_hook(model, self._classifier_hook, layer)
 
-
+@check_noop
 def watch(
     trainer: Trainer,
-    classifier_layer: Union[str, Module] = "classifier",
+    last_hidden_state_layer: Layer = None,
     embedding_dim: Optional[DimensionSlice] = None,
     logits_dim: Optional[DimensionSlice] = None,
-    embedding_fn: Callable = None,
-    logits_fn: Callable = None,
+    classifier_layer: Layer = "classifier",
+    embedding_fn: Optional[Callable] = None,
+    logits_fn: Optional[Callable] = None,
 ) -> None:
-    # store = {}
-    # patch_dataloaders(store)
-    # The columns needed for the forward process
-    signature_cols = add_id_to_signature_columns(trainer)
+    """
+    [`watch`] is used to hook into to the trainer
+    to log to [Galileo](https://www.rungalileo.io/)
+    :param trainer: Trainer object
+    :return: None
+    """
+    a.log_function("transformers_trainer/watch")
 
     # Callback which we add to the trainer
     dqcallback = DQCallback(
-        classifier_layer=classifier_layer,
+        last_hidden_state_layer=last_hidden_state_layer,
         embedding_dim=embedding_dim,
         logits_dim=logits_dim,
+        classifier_layer=classifier_layer,
         embedding_fn=embedding_fn,
         logits_fn=logits_fn,
+    )
+    # The columns needed for the forward process
+    signature_cols = add_id_to_signature_columns(trainer)
+
+    assert trainer.args.dataloader_num_workers == 0, GalileoException(
+        "Parallel Dataloader workers are not supported."
+        "TrainingArgs.dataloader_num_workers should be set to 0"
     )
     # We wrap the data collator to remove the id column
     trainer.data_collator = remove_id_collate_fn_wrapper(
         trainer.data_collator, signature_cols, dqcallback.helper_data
     )
     trainer.add_callback(dqcallback)
-
-
-class BatchSamplerWithLogging(Sampler[List[int]]):
-    r"""Wraps another sampler to yield a mini-batch of indices.
-
-    Args:
-        sampler (Sampler or Iterable): Base sampler. Can be any iterable object
-        batch_size (int): Size of mini-batch.
-        drop_last (bool): If ``True``, the sampler will drop the last batch if
-            its size would be less than ``batch_size``
-
-    Example:
-        >>> list(BatchSampler(SequentialSampler(range(10)),
-                batch_size=3, drop_last=False))
-        [[0, 1, 2], [3, 4, 5], [6, 7, 8], [9]]
-        >>> list(BatchSampler(SequentialSampler(range(10)), batch_size=3,
-        drop_last=True))
-        [[0, 1, 2], [3, 4, 5], [6, 7, 8]]
-    """
-
-    # batch_sampler = BatchSamplerWithLogging(
-    #   RandomSampler(torch.arange(len(a))), 3, False, store)
-
-    def __init__(
-        self,
-        sampler: Union[Sampler[int], Iterable[int]],
-        batch_size: int,
-        drop_last: bool,
-        store: Dict,
-    ) -> None:
-        # Since collections.abc.Iterable does not check for `__getitem__`, which
-        # is one way for an object to be an iterable, we don't do an `isinstance`
-        # check here.
-        if (
-            not isinstance(batch_size, int)
-            or isinstance(batch_size, bool)
-            or batch_size <= 0
-        ):
-            raise ValueError(
-                "batch_size should be a positive integer value, "
-                "but got batch_size={}".format(batch_size)
-            )
-        if not isinstance(drop_last, bool):
-            raise ValueError(
-                "drop_last should be a boolean value, but got "
-                "drop_last={}".format(drop_last)
-            )
-        self.sampler = sampler
-        self.batch_size = batch_size
-        self.drop_last = drop_last
-        self.store = store
-
-    def __iter__(self) -> Iterator[List[int]]:
-        # Implemented based on the benchmarking in:
-        # https://github.com/pytorch/pytorch/pull/76951
-        if self.drop_last:
-            sampler_iter = iter(self.sampler)
-            while True:
-                try:
-                    batch = [next(sampler_iter) for _ in range(self.batch_size)]
-                    self.store["sampler_ids"].append(batch)
-                    self.store["sampler_ids"].append(batch)
-                    yield batch
-                except StopIteration:
-                    break
-        else:
-            batch = [0] * self.batch_size
-            idx_in_batch = 0
-            for idx in self.sampler:
-                batch[idx_in_batch] = idx
-                idx_in_batch += 1
-                if idx_in_batch == self.batch_size:
-                    self.store["sampler_ids"].append(batch)
-                    yield batch
-                    idx_in_batch = 0
-                    batch = [0] * self.batch_size
-            if idx_in_batch > 0:
-                self.store["sampler_ids"].append(batch)
-                yield batch[:idx_in_batch]
